requirements.txt
requests>=2.25.1
schedule>=1.1.0
PyYAML>=5.4.1